+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [unreleased]
 
 - Fix: Update flexbox on `ProjectCreatorCard`.
+- Fix: Add `className` prop on the link `LinkBox`.
 - Feature: Handle breakpoints on the `k-u-blockAlign` utility class.
 
 ## [14.1.4] - 2017-09-22